pyproject.toml
[tool.poetry]
name = "djangorestframework-coursework"
version = "0.1.0"
description = ""
authors = ["Polyaa <polinasatrajkina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.0"
django = "^5.1.7"
djangorestframework = "^3.15.2"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
